+++ app/(tabs)/community/SearchScreen.tsx
@@ -3,12 +3,12 @@ import { addBookmark, removeBookmark } from '@/api/community/bookmarks';
 import CategoryChips, { Category } from '@/components/CategoryChips';
 import Icon from '@/components/common/Icon';
 import ProfileSetupModal from '@/components/common/ProfileSetupModal';
-import PostCard from '@/components/PostCard';
 import SortTabs, { SortKey } from '@/components/SortTabs';
 import WriteFab from '@/components/WriteFab';
 import useMyProfile from '@/hooks/queries/useMyProfile';
 import { useSearchPosts, type PostExFromSearch } from '@/hooks/queries/useSearchPosts';
 import { CATEGORY_TO_BOARD_ID } from '@/lib/community/constants';
+import PostCard from '@/src/features/community/components/PostCard';
 import { useToggleLike } from '@/src/features/community/hooks/useToggleLike';
 import { SearchedPostEx, SearchedPosts, SearchedPostsResp } from '@/src/features/community/types/searchedPostsType';
 import { usePostUI } from '@/src/store/usePostUI';
