+++ app/(tabs)/community/index.tsx
@@ -3,10 +3,10 @@ import { addBookmark, removeBookmark } from '@/api/community/bookmarks';
 import CategoryChips, { Category } from '@/components/CategoryChips';
 import Icon from '@/components/common/Icon';
 import ProfileSetupModal from '@/components/common/ProfileSetupModal';
-import PostCard from '@/components/PostCard';
 import SortTabs from '@/components/SortTabs';
 import WriteFab from '@/components/WriteFab';
 import { CATEGORY_TO_BOARD_ID } from '@/lib/community/constants';
+import PostCard from '@/src/features/community/components/PostCard';
 import { useGetPosts } from '@/src/features/community/hooks/useGetPosts';
 import { useToggleLike } from '@/src/features/community/hooks/useToggleLike';
 import { PostEx, PostsListItem, RequestPageParams } from '@/src/features/community/types/postsListType';
