+++ hooks/queries/useSearchPosts.ts
@@ -1,5 +1,5 @@
 import api from '@/api/axiosInstance';
-import type { Post } from '@/components/PostCard';
+import type { Post } from '@/src/features/community/components/PostCard';
 import { useQuery } from '@tanstack/react-query';
 
 export type SearchPostItem = {
